Moved all runtime webview delegation out of bridge view controller. Removed CAPMessageHandlerWrapper, added WebViewDelegationHandler, renamed CAPAssetHandler.
